# Improve the performance  of file sorter.