<!--
Thank you for contributing to TiDB-CDC! Please read MD's [CONTRIBUTING](https://github.com/pingcap/tidb-cdc/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
#696 

### What is changed and how it works?
In current file sorter, cdc stores `PolymorphicEvent` into file with gob encoding, in the binary format as follows
```text
0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1 2 3
+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
|  data length  |    gob encoding data  |  data length  |
```
Function `readPolymorphicEvent` is reading data length from file and parsing `gob encoding data` data to `model.PolymorphicEvent` once a call.  
In this pr, I changed serialization/deserialization protocol from `gob` to `message pack`.
- [benchmark](https://github.com/qinggniq/ticdc/blob/qinggniq/cdc/puller/filesort_test.go)
  - number of events = 10000
  - size of each event ~= 1024
  - `BenchmarkReadBufXXX` bench *read data length and read gob encoding data without parsing* by `bufio.Buffer` with `4K, 8K, 512K, 4M, 16M` size buffer and `mmap reader`.
  - `BenchmarkGobParseBufXXX` bench *read data length, and parse  gob encoding data to `PolymorphicEvent`* by `bufio.Buffer` of `4K, 8K, 512K, 4M, 16M` size buffer and `mmap reader`.
  - `BenchmarkMsgpackParseBufXXX` bench *read length and parse message pack encoding data to `PolymorphicEvent` using `[message pack](github.com/vmihailenco/msgpack/v5)`* by `bufio.Buffer` of `4K, 8K, 512K, 4M, 16M` size buffer and `mmap reader`.
```shell
[user]$ go test -bench=Benchmark . 
goos: linux
goarch: amd64
pkg: github.com/pingcap/ticdc/cdc/puller
BenchmarkReadBuf4KSingle-16              	     908	   1392144 ns/op
BenchmarkReadBuf8KSingle-16              	     877	   1331969 ns/op
BenchmarkReadBuf512KSingle-16            	     864	   1403036 ns/op
BenchmarkReadBuf4MSingle-16              	     477	   2493812 ns/op
BenchmarkReadBuf16MSingle-16             	     224	   5382727 ns/op
BenchmarkReadBufMmapSingle-16            	     819	   1447675 ns/op
BenchmarkGobParseBuf4KSingle-16          	       8	 141129546 ns/op
BenchmarkGobParseBuf8KSingle-16          	       8	 140973908 ns/op
BenchmarkGobParseBuf512KSingle-16        	       8	 139672640 ns/op
BenchmarkGobParseBuf4MSingle-16          	       8	 137991305 ns/op
BenchmarkGobParseBuf16MSingle-16         	       8	 138779147 ns/op
BenchmarkGobParseBufMmapSingle-16        	       8	 140874481 ns/op
BenchmarkMsgpackParseBuf4KSingle-16      	      98	  12373495 ns/op
BenchmarkMsgpackParseBuf8KSingle-16      	     100	  12136363 ns/op
BenchmarkMsgpackParseBuf512KSingle-16    	     100	  12086008 ns/op
BenchmarkMsgpackParseBuf4MSingle-16      	      87	  12833704 ns/op
BenchmarkMsgpackParseBuf16MSingle-16     	      75	  15980367 ns/op
BenchmarkMsgpackParseBufMmapSingle-16    	     100	  12172507 ns/op
PASS
```
Comparison between `BenchmarkReadBufXXX` and `BenchmarkGobParseBuf4XXX` shows that the performance bottleneck of `readPolymorphicEvent` is `decode` rather than IO. So I investigated golang serialization/deserialization [packages](https://github.com/alecthomas/go_serialization_benchmarks), and found [message pack](https://github.com/vmihailenco/msgpack) to be suitable and fast enough for this situation.