@liuzix 2020-08-03
/run-all-tests